actualizacion

+++ ejercicios c#/ligaBetplay/Menus/MenuConsultaDatos.cs
@@ -2,7 +2,7 @@ namespace ligaBetplay.Menus
 {
     public class MenuConsultaDatos
     {
-        public double menuConsultaDatos()
+        public int menuConsultaDatos()
         {
             Console.Clear();
             Console.WriteLine("\n---------------------------------- CONSULTA DE DATOS ----------------------------------\n");
@@ -10,7 +10,7 @@ namespace ligaBetplay.Menus
             Console.WriteLine("1.2 BUCAR LOS DELANTEROS POR EQUIPO");
             Console.WriteLine("1.3 BUCAR CUERPO DE ENTRENADORES POR EQUIPO");
             Console.WriteLine("1.4 REGRESAR AL MENU PRINCIPAL");
-            return Convert.ToDouble(Console.ReadLine());
+            return Convert.ToInt32(Console.ReadLine());
         }
     }
 }
